# Use the go4.org/mem package to reduce allocations.