--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,7 @@ module github.com/creachadair/jtree
 
 go 1.18
 
-require github.com/google/go-cmp v0.5.6
+require (
+	github.com/google/go-cmp v0.5.6
+	go4.org/mem v0.0.0-20220726221520-4f986261bf13
+)