--- a/stream.go
+++ b/stream.go
@@ -37,7 +37,7 @@ type Handler interface {
 	EndArray(loc Anchor) error
 
 	// Begin a new object member, whose key is at loc. The handler is
-	// responsible for unescaping key values (see jtree.Unescape).
+	// responsible for unescaping key values (see jtree.Unquote).
 	BeginMember(loc Anchor) error
 
 	// End the current object member giving the location and type of the token